refactor: extract bootverify input-alignment checks
